Refactor Binder into its own class
