+++ src/main/java/keycontacts/ui/MainWindow.java
@@ -11,8 +11,6 @@ import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
-import javafx.scene.paint.Color;
-import javafx.scene.shape.Arc;
 import javafx.stage.Stage;
 import keycontacts.commons.core.GuiSettings;
 import keycontacts.commons.core.LogsCenter;
@@ -26,12 +24,6 @@ import keycontacts.logic.parser.exceptions.ParseException;
  * a menu bar and space where other JavaFX elements can be placed.
  */
 public class MainWindow extends UiPart<Stage> {
-
-    // TODO Abstract out the binder into its own class
-    private static final double ARC_RADIUS_X = 8.0;
-    private static final double ARC_RADIUS_Y = 6.0;
-    private static final double ARC_LENGTH = 260.0;
-    private static final double ARC_START_ANGLE = -40.0;
 
     private static final String FXML = "Notebook.fxml";
 
@@ -52,7 +44,7 @@ public class MainWindow extends UiPart<Stage> {
     private AnchorPane rightPage;
 
     @FXML
-    private StackPane binder;
+    private StackPane binderPlaceholder;
 
     @FXML
     private VBox loopParent;
@@ -94,35 +86,11 @@ public class MainWindow extends UiPart<Stage> {
 
         primaryStage.getScene().widthProperty().addListener((observable, oldValue, newValue) -> {
             double newWidth = newValue.doubleValue();
-            leftPage.setPrefWidth((newWidth - binder.getWidth()) / 2);
-            rightPage.setPrefWidth((newWidth - binder.getWidth()) / 2);
+            leftPage.setPrefWidth((newWidth - binderPlaceholder.getWidth()) / 2);
+            rightPage.setPrefWidth((newWidth - binderPlaceholder.getWidth()) / 2);
         });
 
-        primaryStage.getScene().heightProperty().addListener((observable, oldValue, newValue) -> {
-            updateArcs((Double) newValue);
-        });
-    }
 
-    private void updateArcs(double height) {
-        loopParent.getChildren().clear();
-
-        int numArcs = (int) (height / (ARC_RADIUS_Y * 2 + loopParent.getSpacing())) - 5;
-
-        addArcs(numArcs);
-    }
-
-    private void addArcs(int numArcs) {
-        for (int i = 0; i < numArcs; i++) {
-            Arc arc = new Arc();
-            arc.setRadiusX(ARC_RADIUS_X);
-            arc.setRadiusY(ARC_RADIUS_Y);
-            arc.setLength(ARC_LENGTH);
-            arc.setStartAngle(ARC_START_ANGLE);
-            arc.setFill(Color.TRANSPARENT);
-            arc.setStroke(Color.BLACK);
-
-            loopParent.getChildren().add(arc);
-        }
     }
 
     public Stage getPrimaryStage() {
@@ -181,6 +149,10 @@ public class MainWindow extends UiPart<Stage> {
 
         CalendarView calendarView = new CalendarView();
         calendarDisplayPlaceholder.getChildren().add(calendarView.getRoot());
+
+        Binder binder = new Binder(primaryStage.getScene().getHeight());
+        binderPlaceholder.getChildren().add(binder.getRoot());
+        binder.addListener();
     }
 
     /**
