+++ src/main/resources/view/Binder.fxml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.effect.DropShadow?>
+<StackPane xmlns="http://javafx.com/javafx"
+           xmlns:fx="http://javafx.com/fxml">
+    <Pane prefWidth="10.0"
+          style="-fx-border-color: black; -fx-border-width: 0 1 0 1; -fx-background-color: whitesmoke;">
+        <StackPane.margin>
+            <Insets left="5.0" right="5.0"/>
+        </StackPane.margin>
+        <effect>
+            <DropShadow radius="6.335" spread="0.16" width="6.34"/>
+        </effect>
+    </Pane>
+    <VBox fx:id="loopParent" alignment="TOP_CENTER" spacing="5.0"/>
+</StackPane>
