+++ src/main/resources/view/DarkTheme.css
@@ -8,6 +8,11 @@
     -fx-font-size: 14px;
 }
 
+.label-hour {
+    -fx-text-fill: white;
+    -fx-font-family: "Segoe UI Semibold";
+    -fx-font-size: 8px;
+}
 .list-view {
     -fx-background-insets: 0;
     -fx-padding: 0;
